raspberry-pi/requirements.txt
pyserial==3.5
paho-mqtt==2.1.0
python-dotenv==1.1.0
